Fix ensureExchangesAreReleased in CCS tests
